+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
 - Do not initialize default content for Favorites panel [#3346](https://github.com/rokwire/illinois-app/issues/3346).
+- Preserve Home and Browse tout images aspect ratio while loading.
 
 ## [5.0.17] - 2023-06-30
 ### Changed